extracted location manager delegate
