+++ Turfy/Turfy/MapViewController.swift
@@ -88,25 +88,3 @@ class MapViewController: UIViewController {
 	}
     
 }
-
-extension MapViewController: CLLocationManagerDelegate {
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]){
-        let location: CLLocation = locations.first!
-        self.map.centerCoordinate = location.coordinate
-        let reg = MKCoordinateRegionMakeWithDistance(location.coordinate, 1500, 1500)
-        self.map.setRegion(reg, animated: true)
-        geoCode(location: location)
-    }
-    func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus) {
-        map.showsUserLocation = (status == .authorizedAlways)
-    }
-    //Prints error when monitoring fails
-    func locationManager(_ manager: CLLocationManager, monitoringDidFailFor region: CLRegion?, withError error: Error) {
-        print("Monitoring failed for region with identifier: \(region!.identifier)")
-    }
-    //Prints error when locationManager fails
-    func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
-        print("Location Manager failed with the following error: \(error)")
-    }
-}
-
