+++ Turfy/Turfy/VCMapLocationManager.swift
@@ -1,0 +1,34 @@
+//
+//  VCMapLocationManager.swift
+//  Turfy
+//
+//  Created by dev on 29/10/2016.
+//  Copyright © 2016 Lawrence Dawson. All rights reserved.
+//
+
+import Foundation
+import CoreLocation
+import MapKit
+
+extension MapViewController: CLLocationManagerDelegate {
+    
+    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]){
+        let location: CLLocation = locations.first!
+        self.map.centerCoordinate = location.coordinate
+        let reg = MKCoordinateRegionMakeWithDistance(location.coordinate, 1500, 1500)
+        self.map.setRegion(reg, animated: true)
+        geoCode(location: location)
+    }
+    func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus) {
+        map.showsUserLocation = (status == .authorizedAlways)
+    }
+    //Prints error when monitoring fails
+    func locationManager(_ manager: CLLocationManager, monitoringDidFailFor region: CLRegion?, withError error: Error) {
+        print("Monitoring failed for region with identifier: \(region!.identifier)")
+    }
+    //Prints error when locationManager fails
+    func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
+        print("Location Manager failed with the following error: \(error)")
+    }
+}
+
